Move all intergration tests into corresponding location
